refactor: DIActivityTest 테스트 구조 정리
