+++ app/src/test/java/woowacourse/shopping/di/FakeObj.kt
@@ -1,0 +1,3 @@
+package woowacourse.shopping.di
+
+object FakeObj
