--- app/src/test/java/woowacourse/shopping/di/annotationdi/FakeAnnotationDIActivity.kt
@@ -1,7 +1,0 @@
-package woowacourse.shopping.di.annotationdi
-
-import com.di.berdi.DIActivity
-
-class FakeAnnotationDIActivity : DIActivity() {
-    lateinit var viewModel: FakeAnnotationDIViewModel
-}
