--- app/src/test/java/woowacourse/shopping/di/annotationdi/FakeAnnotationDIModule.kt
@@ -1,7 +1,0 @@
-package woowacourse.shopping.di.annotationdi
-
-import com.di.berdi.Module
-
-object FakeAnnotationDIModule : Module {
-    fun provideFakeObj(): FakeObj = FakeObj
-}
