--- app/src/test/java/woowacourse/shopping/di/annotationdi/FakeAnnotationDIViewModel.kt
@@ -1,9 +1,0 @@
-package woowacourse.shopping.di.annotationdi
-
-import androidx.lifecycle.ViewModel
-import com.di.berdi.annotation.Inject
-
-class FakeAnnotationDIViewModel : ViewModel() {
-    @Inject
-    lateinit var fakeObj: FakeObj
-}
