--- app/src/test/java/woowacourse/shopping/di/constructordi/FakeConstructorDIActivity.kt
@@ -1,7 +1,0 @@
-package woowacourse.shopping.di.constructordi
-
-import com.di.berdi.DIActivity
-
-class FakeConstructorDIActivity : DIActivity() {
-    lateinit var viewModel: FakeConstructorDIViewModel
-}
