+++ app/src/test/java/woowacourse/shopping/di/diactivity/FakeObjModule.kt
@@ -1,0 +1,8 @@
+package woowacourse.shopping.di.diactivity
+
+import com.di.berdi.Module
+import woowacourse.shopping.di.FakeObj
+
+object FakeObjModule : Module {
+    fun provideFakeObj(): FakeObj = FakeObj
+}
